Server/requirements.txt
langgraph
langchain
langchain-openai
pandas
python-dotenv
plotly
numpy
flask
flask_cors